Add Food Items hooks and function for centralized axios requests
